% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falconnr.R
\name{cleanup}
\alias{cleanup}
\title{Removes variables and objects loaded from Rcpp modules}
\usage{
cleanup()
}
\description{
Removes variables and objects loaded from Rcpp modules
}
\seealso{
\code{\link{falconnr::cleanup_modules}}
}

