% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falconnr.R
\name{cleanup_modules}
\alias{cleanup_modules}
\title{Removes variables loaded from Rcpp modules}
\usage{
cleanup_modules(pattern = "^Rcpp_")
}
\arguments{
\item{pattern}{regex matching class of objects to remove.
Default: match "Rcpp_" at the beginning
of the class name}
}
\value{
List of names for removed objects
}
\description{
Because the underlying module pointers exposed by Rcpp
do not persist between sessions, is is useful to clean
them up before ending the session, lest a later reference
cause a crash.
}
\details{
This function removes all variables in the calling environment
whose class names, as returned by \code{class()}, match
the given pattern (\code{"^Rcpp_"} by default).
}

