% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppDoc.R
\name{cppDoc.LshNnTable}
\alias{cppDoc.LshNnTable}
\title{\code{LshNnTable}}
\usage{
cppDoc.LshNnTable()
}
\arguments{
\item{tDataMatrix}{-- the data, an R numeric matrix where
each \emph{column} is a data point.
(Thus, pass the transpose of a typical data matrix.)}

\item{params}{-- the LSH configuration parameters


\code{LshNnTable$dimension}}

\item{num_probes}{-- the number of probes to use}

\item{num_candidates}{-- the maximum number of candidates}

\item{q}{-- query point, an R numeric vector of dimension d}

\item{q}{-- query point, an R numeric vector of dimension d}

\item{k}{-- the number of nearest neighbors to return}

\item{q}{-- query point, an R numeric vector of dimension d}

\item{radius}{-- radius around the query point in which to search}

\item{q}{-- query point, an R numeric vector of dimension d}

\item{q}{-- query point, an R numeric vector of dimension d}

\item{queries}{-- matrix of query points, one point per
\emph{column} (pass transpose if necessary)}

\item{answers}{-- vector of indices into data table representing
the nearest neighbor to the corresponding query}

\item{target_precision}{-- minimal correctness probability to achieve}

\item{init_num_probes}{-- number of probes at which to start search}

\item{max_iterations}{-- maximum number of iterations to consider;
default of -1 means no limit}
}
\value{
the dimension of points in the data matrix


\code{LshNnTable$size}

the number points in the data matrix


\code{LshNnTable$getNumProbes}
Returns the number of probes currently being used for multi-probe LSH

the number of probes


\code{LshNnTable$getMaxNumCandidates}
Returns the maximum number of candidates considered for each search

the maximum number of candidates


\code{LshNnTable$setNumProbes}
Set the number of probes used in multi-probe LSH

Note that this is not a costly operation and can be
done repeatedly once the table has been constructed.

a reference to the table object to enable chaining


\code{LshNnTable$setMaxNumCandidates}
Set the maximum number of candidates to consider during similarity search

a reference to the table object to enable chaining

\code{LshNnTable$find_nearest_neighbor}
Find the data point nearest to the given query point

Searches in the encapsulated data set with Locality-Sensitive Hashing.
This is an approximate nearest neighbor search.

index of the data point nearest to \code{q}, wrapped in
        an R integer vector


\code{LshNnTable$find_k_nearest_neighbors}
Find the data points nearest to the given query point

Searches in the encapsulated data set with Locality-Sensitive Hashing.
This is an approximate nearest neighbor search.

vector of indicies of the data points nearest to \code{q},
        wrapped in an R integer vector


\code{LshNnTable$find_near_neighbors}
Find the data points within a specified radius of the given query point

Searches in the encapsulated data set with Locality-Sensitive Hashing.
This is an approximate nearest neighbor search.

vector of indicies of the data points within radius of \code{q},
        wrapped in an R integer vector


\code{LshNnTable$get_candidates}
Find all data points found in a single probing sequence

This is a low-level operation. Note that a single data point might
appear multiple times in the result if several hash tables are used.
These duplicates are included in the returned value.

vector of indicies of the data points found in the search,
        which may include duplicates, wrapped in an R integer vector


\code{LshNnTable$get_unique_candidates}
Find all data points found in a single probing sequence, duplicates removed

This is a low-level operation. Note that a single data point might
appear multiple times in the result if several hash tables are used.
These duplicates are removed from the returned value.

vector of indicies of the data points found in the search,
        which will not include duplicates, wrapped in an R integer vector


\code{LshNnTable$guneNumProbes}
Find number of probes to achieve target precision on training data

During the search for the optimal number of probes, this changes
the set number of probes via \code{setNumProbes()}. However, the
number originally set on entry is restored on exit.

number of probes to use to achieve
}
\description{
Constructor: create a LSH search table for a specified data set
}
\details{
Note: The FALCONN LSH nearest-neighbor search uses a static table,
as contructed here. To add to or change the data set, a new
object needs to be constructed.
}

