% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cppDoc.R
\name{cppDoc.LshParameterSetter}
\alias{cppDoc.LshParameterSetter}
\title{\code{LshParameterSetter}}
\usage{
cppDoc.LshParameterSetter()
}
\arguments{
\item{n}{-- number of data points}

\item{d}{-- dimension of the data points


\code{LshParameterSetter$params}
Returns a copy of the underlying parameters structure}

\item{distance}{-- one of the strings: "negative_inner_product", 
"euclidean_squared", or "unknown"; all other values
lead to a setting of "unknown".}

\item{distance}{-- one of the strings: "negative_inner_product", 
"euclidean_squared", or "unknown"; all other values
lead to a setting of "unknown".}

\item{tables}{-- a positive integer, the number of hash functions
to use in locality-sensitive hashing}

\item{tables}{-- a positive integer, the number of hash tables
to use in locality-sensitive hashing}

\item{storage}{-- one of the strings: "flat_hash_table",
"bit_packed_flat_hash_table", "stl_hash_table",
"linear_probing_hash_table", or "unknown"; 
all other values lead to a setting of "unknown".}

\item{family}{-- one of the strings: "hyperplane", "cross_polytope",
or "unknown"; all other values lead to a setting
of "unknown".}

\item{rotations}{-- number of rotations to use
or "unknown"; all other values lead to a setting
of "unknown".}
}
\value{
instance of \code{struct LSHConstructionParameters} representing
        the underlying LSH parameters


\code{LshParameterSetter$withDefaults}
Sets all parameters to their default values based on data size

a reference to the original object, enabling chaining


\code{LshParameterSetter$distance}
Sets the distance function used in similarity search

a reference to the original object, enabling chaining


\code{LshParameterSetter$numHashFunctions}
Sets the number of hash functions used in locality-sensitive hashing

a reference to the original object, enabling chaining


\code{LshParameterSetter$numHashTables}
Sets the number of hash tables used in locality-sensitive hashing

a reference to the original object, enabling chaining


\code{LshParameterSetter$storage}
Sets the storage mode used in similarity search

a reference to the original object, enabling chaining


\code{LshParameterSetter$family}
Sets the family used for locality-sensitive hashing

a reference to the original object, enabling chaining


\code{LshParameterSetter$rotations}
Sets the number of pseudo-rotations used with cross-polytope hash

For sparse data, a value of 2 is recommended; for dense data, 1.

a reference to the original object, enabling chaining


\code{LshParameterSetter$asList}
Represents parameters as an R list

R-list with names corresponding to the parameters
}
\description{
Constructor: create an instance of an \code{LshParameterSetter}
}
\seealso{
\code{\link{falconn/lsh_nn_table.h}}, specifically
\code{struct LSHConstructionParameters} for details of the returned 
structure.
}

