% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/falconnr.R
\docType{package}
\name{falconnr}
\alias{falconnr}
\alias{falconnr-package}
\title{falconnr: A package for locality-sensitive hashing nearest-neighbor search}
\description{
The falconnr package provides an interface to the
\href{https://falconn-lib.org/}{FALCONN} C++ library
that supports similarity-search for high-dinensional data
based on locality-sensitive hashing.
}
\author{
Christopher R. Genovese \email{genovese@cmu.edu}
}
\references{
\url{https://falconn-lib.org/}
}

