% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh.R
\docType{methods}
\name{initialize,LshTable-method}
\alias{initialize,LshTable-method}
\title{Initialize a LshTable given a data matrix}
\usage{
\S4method{initialize}{LshTable}(.Object, X, params = NULL)
}
\arguments{
\item{.Object}{-- the LshTable object to be initialized}

\item{X}{-- the data, a matrix with each \emph{row}
                  corresponding to a point.

Note the different orientation of \code{X} relative to the
interface of the Rcpp-exposed constructor of \code{LshNnTable},
which expects each \emph{column} to be a point.}
}
\description{
Initialize a LshTable given a data matrix
}

