% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh.R
\docType{methods}
\name{similar,LshTable-method}
\alias{similar,LshTable-method}
\title{Search for data points close to the given query point}
\usage{
\S4method{similar}{LshTable}(object, query, k = 1, radius = NULL,
  points = FALSE)
}
\arguments{
\item{object}{-- an LshTable object}

\item{query}{-- a vector whose dimension matches the dimension
of the data in object}

\item{k}{-- number of nearest neighbors to find,
ignored if \code{radius} is supplied}

\item{radius}{-- a threshold for near-neighbor search}

\item{points}{-- if FALSE, return indices in the data matrix of found points,
otherwise, return a submatrix of the found points}
}
\description{
Search for data points close to the given query point
}

