% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsh.R
\name{similar}
\alias{similar}
\title{Search for data points similar to a given query point}
\usage{
similar(object, query, ...)
}
\arguments{
\item{object}{-- an object representing data to search}

\item{query}{-- a vector whose dimension matches the dimension
of the data in object}
}
\description{
Search for data points similar to a given query point
}

